feat: added terms and privacy pages
