+++ data/team.json
@@ -24,7 +24,7 @@
     {
         "name": "Adamantios Zaras",
         "imageSrc": "adamantion.jpg",
-        "position": "Python Engineer - Coordinator"
+        "position": "Senior Python Engineer - Coordinator"
     },
     {
         "name": "Adriano",
@@ -49,7 +49,7 @@
     {
         "name": "Ardian",
         "imageSrc": "Ardian.png",
-        "position": "Software Engineer"
+        "position": "Senior Software Engineer"
     },
     {
         "name": "Artem",
@@ -109,7 +109,7 @@
     },
     {
         "name": "Mohan Das",
-        "position": "Frontend Engineer"
+        "position": "Senior Frontend Engineer"
     },
     {
         "name": "Ojuswi Rastogi",
